requirements.txt
Flask==2.3.2
greenlet==2.0.2
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
pytest==7.3.1
astral==3.2
pytz==2023.3
PyYAML==6.0
geopy==2.3.0
